% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asarray.R
\name{AsArray}
\alias{AsArray}
\title{Convert objects to arrays}
\usage{
AsArray(x)
}
\arguments{
\item{x}{Any object}
}
\value{
The same object with all attributes removed except for dim and dimnames
}
\description{
Convert objects to arrays
}
