% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.operators.R
\name{CheckInputVariableLabelsChanged}
\alias{CheckInputVariableLabelsChanged}
\title{Inspect two variable sets and both share the same variable names}
\usage{
CheckInputVariableLabelsChanged(input, original.variable.labels, function.name)
}
\arguments{
\item{input}{A list with two variable sets (data.frames) to check the names}

\item{original.variable.labels}{A list containing the variable names of the original inputs}

\item{function.name}{String containing the function name, used in the thrown error messages.}
}
\value{
The (possibly modified) data.frames after they have been validated
}
\description{
Inspect two variable sets and both share the same variable names
}
\details{
Takes two inputs and checks if both variable set inputs contain the same variable names
 on their columns (after ignoring data reductions).
 These can be of different order but all elements need to exist in both inputs.
 If that isn't the case then the function will thrown an error with a message containing the
 variable names that are not present in both inputs.
}
