% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.operators.R
\name{CheckInputVariableNamesMatch}
\alias{CheckInputVariableNamesMatch}
\title{Inspect two variable sets and both share the same variable names}
\usage{
CheckInputVariableNamesMatch(
  first,
  second,
  remove.columns = c("NET", "SUM", "Total"),
  function.name
)
}
\arguments{
\item{first}{First Variable Set to perform the matching}

\item{second}{Second Variable Set to perform the matching}

\item{remove.columns}{character vector containing variable names (column names) to
ignore}

\item{function.name}{String containing the function name, used in the thrown error messages.}
}
\description{
Inspect two variable sets and both share the same variable names
}
\details{
Takes two inputs and checks if both variable set inputs contain the same variable names
 on their columns (after ignoring data reductions).
 These can be of different order but all elements need to exist in both inputs.
 If that isn't the case then the function will thrown an error with a message containing the
 variable names that are not present in both inputs.
}
