% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{checkSelections}
\alias{checkSelections}
\alias{checkSelections.character}
\title{Check indices are valid for selecting from a table}
\usage{
checkSelections(indices, ...)

\method{checkSelections}{character}(indices, table, dim, ...)
}
\arguments{
\item{indices}{A vector to be used to select from a table}

\item{...}{Currently, ignored}

\item{table}{A table (matrix, data.frame, etc.) to check \code{indices} against.}

\item{dim}{Integer specficying the dimension of \code{table} to consider.}
}
\value{
A possibly modified copy of \code{indices} with invalid
    values removed.
}
\description{
Validates a vector of indices to be used to select from a dimension
of a table; e.g., removing duplicates and indices out of range for
the supplied table with a warning.
}
