% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{DropMultipleStatisticsFromTable}
\alias{DropMultipleStatisticsFromTable}
\title{Select the first statistic from a Q Table}
\usage{
DropMultipleStatisticsFromTable(
  table,
  warn = TRUE,
  drop.stats.from.2d.table = FALSE
)
}
\arguments{
\item{table}{A Q Table (an \code{array} of 1 to 5 dimensions)
containing multiple statistics (which vary along the final
dimension)}

\item{warn}{Logical; should a warning be thrown when dropping
statistics?}

\item{drop.stats.from.2d.table}{Logical; should multiple statistics
be dropped from 2D Q Tables? Defaults to \code{FALSE} for the
Table - Select feature in Displayr.}
}
\value{
\code{table} with only the first element (statistic) from
    the last dimension
}
\description{
Modifies a Q Table input containing multiple statistics, to an
identical Q Table containing only the first statistic.
}
