% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstlast.R
\name{First}
\alias{First}
\title{First elements in an object}
\usage{
First(x, keep = 6L, by = "element", calendar = TRUE, ...)
}
\arguments{
\item{x}{An object such as a \code{vector}, \code{matrix}, \code{array}, \code{list} etc.}

\item{keep}{An \code{integer} of the number of first elements to keep, or
if \code{by} is a unit of time, the number initial time periods to
consider when retaining elements.}

\item{by}{A string that is either \code{"element"} or a unit of time:
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"},
\code{"day"}, \code{"hour"}, \code{"minute"}, \code{"second"}.}

\item{calendar}{Whether to consider calendar periods, e.g., with calendar months,
actual months are counted, i.e., January, February etc. Otherwise months are
counted as time periods of 28 to 31 days (based on the calendar months
between start and end dates).}
}
\value{
A subset containing the first elements in x
}
\description{
Return first elements in an object, when the number of
  initial elements can be defined in terms of the number of elements, or
  if the data is named with dates, the first elements within a specified
  number of time periods.
}
\examples{
  First(1:10, 6) # 1:6
  x <- 1:10
  names(x) <- Sys.Date() + 1:10
  First(x, keep = 1, by = "week", calendar = FALSE) # next 7 days
}
