% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstlast.R
\name{First}
\alias{First}
\alias{Last}
\title{First and last elements in an object}
\usage{
First(x, keep = 6L, by = "element", calendar = TRUE, ...)

Last(x, keep = 6L, by = "element", calendar = TRUE, ...)
}
\arguments{
\item{x}{An object such as a \code{vector}, \code{matrix}, \code{array}, \code{list} etc.}

\item{keep}{A scalar or vector of the number of first elements to keep, or
if \code{by} is a unit of time, the number initial time periods to
consider when retaining elements. If \code{x} is multi-dimensional,
e.g. a matrix, \code{keep} can be a vector whose entries correspond to
the dimensions, which allows the first or last elements to be kept in
multiple dimensions, e.g. both rows and columns. If all elements in a
dimension are to be kept, the corresponding value for \code{keep} should
be \code{NA}.}

\item{by}{A scalar or vector that contains either \code{"element"}
or a unit of time: \code{"year"}, \code{"quarter"}, \code{"month"},
\code{"week"}, \code{"day"}, \code{"hour"}, \code{"minute"},
\code{"second"}. If \code{keep} is a vector, then \code{by} can also be a
vector with elements corresponding to keep, allowing different dimensions
to have different values specified.}

\item{calendar}{A logical scalar that indicates whether to consider calendar
periods when \code{by} is a unit of time. See details for more information
on calendar and non-calendar periods.}

\item{...}{Additional arguments for \code{head} and \code{tail}.}
}
\value{
A subset containing the first elements in x
}
\description{
Return first elements in an object, when the number of
  initial elements can be defined in terms of the number of elements, or
  if the data is named with dates, the first elements within a specified
  number of time periods.
}
\details{
When calendar periods are considered, a year is counted as the
  12-month period from January to December; a quarter is one of the
  following 3-month periods: January-March, April-June, July-September,
  October-December; a month is the period from the first to last day in the
  12 months in the calendar; a week is the 7 day period from Sunday to
  Saturday; a day is the 24 hour period from midnight to midnight; and hours,
  minutes and seconds are the fixed periods as indicated on a clock. For
  example, if the first 2 calendar years are to be kept and the earliest
  date in the data is Nov 6, 2020, all data labeled in 2020 and 2021 will be
  kept.

  When non-calendar periods are considered instead, all periods are compared
  against a reference date-time, which is the earliest date-time in the
  labels for \code{First}, and the latest date-time in the labels for
  \code{Last}. So for example if the reference date is Nov 6, 2020, and the
  first 2 years of data are to be kept, all data within two years from
  Nov 6, 2020 would be retained.
}
\examples{
  First(1:10, 6) # 1:6
  x <- 1:10
  names(x) <- Sys.Date() + 1:10
  First(x, keep = 1, by = "week", calendar = FALSE) # next 7 days
}
