% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{FlattenQTable}
\alias{FlattenQTable}
\title{Flatten a QTable}
\usage{
FlattenQTable(x, drop = FALSE)
}
\arguments{
\item{x}{A QTable to be flattened}

\item{drop}{If a multi statistic table, should only the first statistic be retained.}
}
\description{
Flatten a QTable
}
