% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{FlattenQTableToMatrix}
\alias{FlattenQTableToMatrix}
\title{Convert High-Dimensional Q Tables to Matrix}
\usage{
FlattenQTableToMatrix(x, row.dims, col.dims)
}
\arguments{
\item{x}{A 4D or 3D Q Table containing a single statistic to be
converted to matrix.}

\item{row.dims}{length one or two integer vector specifying the
dimensions of \code{x} to use for the rows of the output
matrix.}

\item{col.dims}{length one or two integer vector specifying the
dimensions of \code{x} to use for the rows of the output
matrix.}
}
\value{
A matrix (2D array) containing a flattened version of
    \code{x} with multiple variables varying in possibly both
    dimensions, similar to how the table is displayed in Q and
    Displayr.
}
\description{
Flattens/melts a Q Table with more than two dimensions into a
matrix using \code{ftable}, ensuring the output table has
informative  names.
}
\seealso{
\code{ftable}
}
