% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{FlattenTableAndDropStatisticsIfNecessary}
\alias{FlattenTableAndDropStatisticsIfNecessary}
\title{Convert Q Tables to Tabular (2D) Format}
\usage{
FlattenTableAndDropStatisticsIfNecessary(
  table,
  drop.stats.from.2d.table = FALSE
)
}
\arguments{
\item{table}{A Q Table (an array of 1-5 dimensions with special
attributes)}

\item{drop.stats.from.2d.table}{Logical; should multiple statistics
be dropped from 2D Q Tables? Defaults to \code{FALSE} for the
Table - Select feature in Displayr.}
}
\value{
A matrix (2D array) containing a flattened version of
    \code{x} with only its first statistic (unless \code{table} is
    2D and \code{drop.stats.from.2d.table} is \code{FALSE}. It will
    contain multiple variables varying in possibly both dimensions,
    similar to how the table is displayed in Q and Displayr.
}
\description{
Flattens/melts a Q Table of more than two dimensions, dropping
multiple statistics if they are present and returning only the first in
a flattened matrix.
}
