% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.operators.R
\name{GetVariableSetLabels}
\alias{GetVariableSetLabels}
\title{Inspect data reduced labels with the codeframe attribute}
\usage{
GetVariableSetLabels(x)
}
\arguments{
\item{x}{A Variable or VariableSet with codeframe attributes}
}
\value{
A character vector with labels that exist in the code frame or the variable names from the column names
}
\description{
Inspect the codeframe to deduce the appropriate variable labels, if the code frame exists.
Otherwise, just use the names of the data.frame
}
