% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscript-tables.R
\name{IsQTableAttribute}
\alias{IsQTableAttribute}
\title{Check character vector contains Q Table attribute names.}
\usage{
IsQTableAttribute(
  attribute.names,
  qtable.attrs = c("statistic", "dim", "dimnames", "mapped.dimnames", "dimnets",
    "dimduplicates", "span", "basedescriptiontext", "basedescription",
    "QStatisticsTestingInfo", "questiontypes", "footerhtml", "name", "questions")
)
}
\arguments{
\item{attribute.names}{The character vector of attributes to inspect}

\item{qtable.attrs}{Optional character vector to check attribute names against.
Defaults to the know names of attributes for QTables.}
}
\value{
A logical vector if the input character vector contains Q Tabel attribute names
}
\description{
Check character vector contains Q Table attribute names.
}
