% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstlast.R
\name{Last}
\alias{Last}
\title{Last elements in an object}
\usage{
Last(x, keep = 6L, by = "element", calendar = TRUE, ...)
}
\arguments{
\item{x}{An object such as a \code{vector}, \code{matrix}, \code{array}, \code{list} etc.}

\item{keep}{An \code{integer} of the number of last elements to keep, or
if \code{by} is a unit of time, the number final time periods to
consider when retaining elements.}

\item{by}{A string that is either \code{"element"} or a unit of time:
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"},
\code{"day"}, \code{"hour"}, \code{"minute"}, \code{"second"}.}

\item{calendar}{Whether to consider calendar periods, e.g., with calendar months,
actual months are counted, i.e., January, February etc. Otherwise months are
counted as time periods of 28 to 31 days (based on the calendar months
between start and end dates).}

\item{...}{Additional arguments for \code{head} and \code{tail}.}
}
\value{
A subset containing the last elements in x
}
\description{
Return last elements in an object, when the number of
  final elements can be defined in terms of the number of elements, or
  if the data is named with dates, the last elements within a specified
  number of time periods.
}
\examples{
  Last(1:10, 6) # 4:10
  x <- 1:10
  names(x) <- Sys.Date() - 1:10
  Last(x, keep = 1, by = "week", calendar = FALSE) # previous 7 days
}
