% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{ParseCategoricalLabels}
\alias{ParseCategoricalLabels}
\title{Parsing Categorical Labels to Count}
\usage{
ParseCategoricalLabels(concatted.labels, input)
}
\arguments{
\item{concatted.labels}{A single chracter element containing the delimited cagetories to count
e.g. "foo, bar, baz" or "foo; bar; baz"}

\item{input}{Either a \code{data.frame} containing factors or a factor or numeric object.
If any other input is given, then NULL is returned.
\code{concatted.labels}.}
}
\value{
A delimited string containing only the categorical labels that exist in the supplied \code{data.frame}
}
\description{
Checks a delimited string contains the levels of factors contained in a data.frame. If there
   are unmatched elements in the string, then a warning is thrown advising the user of the inability to match
}
