% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{ParseCategoriesToRemove}
\alias{ParseCategoriesToRemove}
\title{Parse Categories to Remove in GUI textboxes}
\usage{
ParseCategoriesToRemove(input.str, inputs)
}
\arguments{
\item{input.str}{A list containing two characters with the desired categories
in each dimension to be removed. The first element for the rows, the second for
the columns. The expected format is either a comma separated string or ;
separated string for each dimension.}

\item{inputs}{A list containing all the inputs that have rows and columns that
are potentially removed.}
}
\value{
A list with two character vectors that contain the category names of
  all row/column slices that are to be removed.
}
\description{
Parse Categories to Remove in GUI textboxes
}
