% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{SelectFromTable}
\alias{SelectFromTable}
\title{Select from row and columns of tables}
\usage{
SelectFromTable(
  table,
  row.selection.mode = c("vector", "first rows", "last rows", "range",
    "first date-time periods", "last date-time periods", "date range"),
  row.selections = NULL,
  column.selection.mode = c("vector", "first columns", "last columns", "range",
    "first date-time periods", "last date-time periods", "date range"),
  column.selections = NULL,
  unit = NULL,
  calendar = TRUE,
  ...
)
}
\arguments{
\item{table}{A table (\code{matrix} or \code{array}) to select
from.}

\item{row.selection.mode}{String specifying how to select rows to
    include in the output table. One of
\enumerate{
\item \code{"vector"} - The default mode, use to specify integer, logical, or
    character vector indices to perform the selection similar to
    \code{[}.
\item \code{"first rows"} - Similar to \code{\link{head}}, use this mode to
select the first \code{n} rows of the table, where \code{n} is supplied
in \code{row.selections}.
\item \code{"last rows"} - Similar to \code{\link{tail}}, use this mode to
select the last rows of the table.
\item \code{"First date-time periods"} - Use this mode if the row names of
the table contain dates to select the first rows of the table within a
certain time period specified by \code{unit} (e.g. within the first three days
from the first date in the rows names of \code{table}).
\item \code{"Last date-time periods"} - For tables containing dates, this
mode allows for selecting the last rows of the table falling in the specified
time period (e.g. within the last three quarters of the final date in the
rows names).
\item \code{"range"} - use this mode to supply a range of indices to select
\item \code{"date range"} - use this mode to supply a range of dates to select
when the rows of \code{table} contains dates.
}}

\item{row.selections}{For \code{row.selection.mode = "vector"} an integer
vector providing the row indices of \code{table} to include in the output, a logical
vector with the same length as rows in the table, or a
character vector containing the row names to include. When
\code{row.selection.mode} is \code{"First rows"}, \code{"Last rows"},
\code{"First date-time periods"}, or \code{"Last date-time periods"}; a
scalar integer specifying the number of first/last rows to include in
the output. When \code{row.selection.mode = "range"}, a string containing a
(possibly comma-separated) list of indices and ranges to include; e.g.
\code{"5-8"} or \code{"1,3-5,9"}. When \code{row.selection.mode = "date range"},
a string containing two dates separated by \code{"--"}; e.g.,
\code{"2002-01-15--2004-06-30"}. Only rows with names containing dates
that lie between the two dates in the range are included in the output.}

\item{column.selection.mode}{as \code{row.selection.mode} except
specifies the mode to use for selecting columns. Possible
values are \code{"vector"}, \code{"first columns"},
\code{"last columns"}, \code{"range"}, \code{"date range"},
\code{"first date-time periods"},
\code{"last date-time periods"}.}

\item{column.selections}{same as \code{row.selections} except for
rows.}

\item{unit}{See \code{\link{First}}, the unit of time to consider
when the selection mode uses dates. Ignored if the selection
mode does not involve dates. Either a scalar or length-2 vector
to use different units in rows and columns.}

\item{calendar}{Only used when selection mode involves dates. A
logical scalar that indicates whether to consider calendar
periods when the selection mode involves dates. Can also be
length-2 vector to use different options in rows and
columns. See Details in \code{\link{First}}.}

\item{...}{Additional arguments passed to \code{\link{First}}.}
}
\value{
A matrix containing the selected rows and columns of the
    table specified by \code{table.to.show}.
}
\description{
Subsets tables supporting a number of different modes for
specifying the rows and columns to retain. In addition to the usual
integer and character vector subsetting similar to \code{[}, first
and last rows and columns can be retained similar to
\code{\link{head}} and \code{\link{tail}}, and ranges of indices or
dates can be specified as a character string. Special support for
selecting from tables with dates in the rows and columns is also
provided.
}
\details{
For Q Tables containing multiple statistics, only the
    first statistic will be shown in the output table (with a
    warning); unless the table contains a single-variable Variable
    Set, which permits selecting statistics using
    \code{column.selections}.

Q Tables that stored as array of more than two dimensions are first
flatten using \code{\link{FlattenTableAndDropStatisticsIfNecessary}}.
}
\examples{
x <- matrix(1:6, 3, 2)
rownames(x) <- LETTERS[1:3]
SelectFromTable(x, row.selection.mode = "vector", row.selections = c("B", "C"))
SelectFromTable(x, row.selection.mode = "Last rows", row.selections = 2)

SelectFromTable(t(1:6), column.selection.mode = "range",
                column.selections = "2-4,6")
}
\seealso{
\code{\link{First}}, \code{\link{Last}},
    \code{\link{FlattenTableAndDropStatisticsIfNecessary}}
}
