% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.operators.R
\name{SingleVariableAsDataFrame}
\alias{SingleVariableAsDataFrame}
\title{Wrap a Variable in a data.frame}
\usage{
SingleVariableAsDataFrame(x)
}
\arguments{
\item{x}{A Variable (vector) that has label or name attribute.}
}
\value{
A data.frame with one column named after the variable label or name attribute.
}
\description{
Wrap a Variable in a data.frame
}
