% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{Sum}
\alias{Sum}
\title{General-Purpose Element-wise Summation}
\usage{
Sum(..., ignore.missing = TRUE)
}
\arguments{
\item{...}{Objects to be summed; e.g. Q Tables, arrays}

\item{ignore.missing}{Should missing values removed? Defaults to \code{TRUE},
differing from \code{na.rm} in \code{\link{sum}}.}
}
\value{
The sum.
}
\description{
A replacement for \code{\link{sum}} supporting element-wise
summation of many objects in addition to vectors.
}
\examples{
Sum(1:3, 2:4)
}
