% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R, R/sumrows.R
\name{Sum}
\alias{Sum}
\alias{SumRows}
\title{General-Purpose Summation functions

A replacement for \code{\link{sum}} supporting element-wise
summation of many objects in addition to vectors.}
\usage{
Sum(
  ...,
  remove.missing = TRUE,
  remove.rows = c("NET", "SUM", "Total"),
  remove.columns = c("NET", "SUM", "Total"),
  subset = NULL,
  weights = NULL,
  warn = FALSE
)

SumRows(
  ...,
  remove.missing = TRUE,
  remove.rows = c("NET", "SUM", "Total"),
  remove.columns = c("NET", "SUM", "Total"),
  subset = NULL,
  weights = NULL,
  match.elements = "Yes - ignore if unmatched",
  warn = FALSE
)
}
\arguments{
\item{...}{Objects to be summed; e.g. Q Tables, arrays, vectors, Variables or Variable Sets.
Depending on if \code{Sum}, \code{SumRows} or \code{SumColumns} is used, different possible
multiple input types are allowed. See the}

\item{remove.missing}{Logical element controlling whether missing values should be
removed in calculations? Defaults to \code{TRUE}. If set to \code{FALSE} and one of
input elements contains missing values, then the resulting computed sum will also be
\code{NA}.}

\item{remove.rows}{The labels of named vectors, or, rows of tables and similar
structures to be removed from the row dimension of the input. Any row elements
with the labels specified here will not be included in the resulting sum calculation.}

\item{remove.columns}{The labels of columns of tables and similar
structures to be removed from the column dimension of the input. Any column elements
with the labels specified here will not be included in the resulting sum calculation.}

\item{subset}{Logical vector of units to use in the calculation. Only applied to variables and
not to QTables that contain statistics since the original variable data is unavailable.}

\item{weights}{Numeric vector of weights to use in the calculation. Only applies to variables
and not to QTables that contain statistics since the original variable data is unavailable.}

\item{warn}{Logical element to control whther warnings are shown when non-obvious
operations are performed (e.g., removing rows, removing missing values when they are present).}

\item{match.elements}{String element specifying if inputs with named row elements
(or named elements for a single vector) should be matched when used in \code{SumRows}.}
}
\value{
The sum or sum of the rows or sum of the columns of the provided inputs respectively
  depending on the function used.
}
\description{
General-Purpose Summation functions

A replacement for \code{\link{sum}} supporting element-wise
summation of many objects in addition to vectors.
}
\details{
TBA
}
\examples{
# Examples using Sum
Sum(1:3, 2:4)
all.equal(Sum(1:3, 2:4), sum(1:3, 2:4))
x <- 1:4
names(x) <- c(LETTERS[1:3], "SUM")
Sum(x)
x <- 1:9
desired.subset = rep(c(TRUE, FALSE), c(3, 6))
Sum(x, subset = desired.subset)
desired.weights <- runif(9)
y <- 10:18
Sum(x, y, weights = desired.weights)
sum(x * desired.weights, y * desired.weights)
basic.table <- matrix(1:12, nrow = 4, dimnames = list(letters[1:4], LETTERS[1:3]))
table.with.marginals <- rbind(basic.table, SumColumns(basic.table))
table.with.marginals <- cbind(table.with.marginals, c(SumRows(basic.table), Sum(basic.table)))
dimnames(table.with.marginals) <- list(c(letters[1:4], "SUM"), c(LETTERS[1:3], "SUM"))
all.equal(Sum(table.with.marginals), sum(basic.table))
}
