% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.operators.R
\name{Divide}
\alias{Divide}
\alias{Multiply}
\alias{Subtract}
\title{General-Purpose Arithmetic functions}
\usage{
Divide(
  numerator = NULL,
  denominator = NULL,
  remove.missing = TRUE,
  remove.rows = NULL,
  remove.columns = NULL,
  match.elements = "Yes",
  subset = NULL,
  warn = FALSE
)

Multiply(
  multiplicand = NULL,
  multiplier = NULL,
  remove.missing = TRUE,
  remove.rows = NULL,
  remove.columns = NULL,
  match.elements = "Yes",
  subset = NULL,
  weights = NULL,
  warn = FALSE
)

Subtract(
  minuend = NULL,
  subtrahend = NULL,
  remove.missing = TRUE,
  remove.rows = NULL,
  remove.columns = NULL,
  match.elements = "Yes",
  subset = NULL,
  weights = NULL,
  warn = FALSE
)
}
\arguments{
\item{numerator}{Numerator in the division e.g. a scalar, vector, matrix, Variable, Variable Set or Q Table}

\item{denominator}{Denominator in the division, assuming to be the same data structure
as the numerator or can be coerced into an equivalent data structure.}

\item{remove.missing}{Logical element controlling whether missing values should be
removed in calculations? Defaults to \code{TRUE}. If set to \code{FALSE} and if one of
input elements contains missing values, then the resulting computed sum will also be
\code{NA}.}

\item{remove.rows}{The labels of named vectors, or, rows of tables and similar
structures to be removed from the row dimension of the input. Any row elements
with the labels specified here will not be included in the resulting \code{Sum}
calculation.}

\item{remove.columns}{The labels of columns of tables and similar
structures to be removed from the column dimension of the input. Any column elements
with the labels specified here will not be included in the resulting \code{Sum}
calculation.}

\item{match.elements}{Either a single string with three possible options or named character vector with two elements. The possible single character options are: \itemize{
\item "No": Ignores names and requires either inputs which the same dimensions
      or partially agreeing where recycling can be performed. See details for more information.
\item "Yes - hide unmatched" or "Yes": Performs a matching algorithm that checks row names and column names
      of all elements and attempts to match the appropriate elements. The matching
      will check both exact matches or fuzzy matches and permutes the order of the elements
      so that the names match. It also may transpose an input if,
      for example, the column names of one input match the row names of another input.
      Any unmatched elements are removed from the input before calculation.
\item "Yes - show unmatched": Performs the same matching algorithm above but any
      unmatched elements are kept in the input. The other input without the element has
      missing values spliced in.
      }
 A named character vector is possible but it must have two elements named \code{rows}
 or \code{columns} (partial matching names permissible) that specify the matching behavior
 for rows and columns.
 E.g. \code{match.elements = c(rows = "Yes - hide unmatched", columns = "No")} to specify
 that rows are to be matched but unmatched rows are to be removed from the calculation.
 The columns are not to be matched. The full set of alternatives for either \code{match.rows}
 or \code{match.columns} are given below but are described only for the row scenario.
\itemize{
  \item \code{"Yes"} or \code{"Yes - hide unmatched"}: Perform an exact name match between
  the row names of input elements. Any unmatched row names will cause the entire row to be
  removed before calculation.
  \item \code{"Yes - show unmatched"} Performs an exact name match between
  input elements in the same manner as \code{"Yes"} option. However, any row names that are
  not matched in the other outputs are permissible. The input that doesn't have that row
  will have a row of missing values spliced in. The resulting row will either be \code{NA} if
  \code{remove.missing} is set to \code{FALSE} or zero if \code{remove.missing} is set
  to \code{TRUE}.
  \item \code{"Fuzzy"} Same behaviour as \code{"Yes"} except the matching uses fuzzy name
  matching. First an exact name match is performed, any remaining unmatched elements are attempted to
  be matched using near matches using the Levenshtein distance. The algorithm will
  match elements that have a one-to-one matching with a Levenshtein distance of
  at most 1 unit, ignoring case. Following this, some common variants of questionnaire
  reponses are attempted to be matched. E.g. \code{"None of these"} could be matched
  with \code{"none"}. This occurs for variants of \code{"Other"}, \code{"All of these"}
  and variants of \code{"Don't know"}. If further unmatched elements exist, punctuation
   and white space are removed and matches attempted.
  \item \code{"Fuzzy - show unmatched"} Similar to \code{"Fuzzy"} except any unmatched row
  names will have a row of missing values spliced in similar to \code{"Yes - show unmatched"}.
  \item \code{"No"} No matching is to occur with the row names. In this case, all
  input elements need to have the same number of rows.
  }}

\item{subset}{Logical vector of units to use in the calculation. Only applied to variables and
not to Q Tables that contain statistics since the original variable data is unavailable.}

\item{warn}{Logical element to control whether warnings are shown when non-obvious
operations are performed (e.g., removing rows, removing missing values when they are present).
Possible warnings presented include \itemize{
  \item Whether missing values were identified and removed if \code{remove.missing}
    is set to \code{TRUE}
  \item Whether different statistics are being summed in the case of Q Table.
  \item Whether unmatched rows or columns have been removed from the calculation if the user
  has specified for unmatched elements to be hidden.
  \item Whether any inputs have been recycled to higher dimensional forms.
}}

\item{weights}{Numeric vector of weights to use in the calculation. It is required to have the same
number of elements as there are rows in the inputs as the weight vector is applied across the row
dimension (elements in different columns but the same row will have the same weight element applied). The
exception to this is that weights will not be applied to Q Tables containing statistics since
the original variable data is unavailable.}
}
\description{
\code{Divide} is a generalization of the / function in \code{\link{Arithmetic}}
  but supporting additional pre-processing and matching of data before calculation.
}
\examples{
# TBA

}
