% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathematical.operators.R
\name{CheckInputVariableLabelsChanged}
\alias{CheckInputVariableLabelsChanged}
\title{Inspect multiple variable sets that should have common variable labels}
\usage{
CheckInputVariableLabelsChanged(input, original.variable.labels, function.name)
}
\arguments{
\item{input}{A list with one or more variable sets (data.frames) to check the labels}

\item{original.variable.labels}{A list containing the variable labels of the original inputs}

\item{function.name}{String containing the function name, used in the thrown error messages.}
}
\value{
The (possibly modified) data.frames after they have been validated
}
\description{
Inspect multiple variable sets that should have common variable labels
}
\details{
Takes a list of multiple inputs and checks if all variable set inputs contain the same variable names
 on their columns (after accounting for data reductions).
 These can be of different order but all elements need to exist in both inputs.
 If that isn't the case then the function will thrown an error with a message containing the
 variable labels that are not present in all inputs.
}
