% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{FlattenQTable}
\alias{FlattenQTable}
\title{Flatten a QTable}
\usage{
FlattenQTable(x, drop = FALSE)
}
\arguments{
\item{x}{A QTable to be flattened}

\item{drop}{If a multi statistic table, should only the first statistic be retained.}
}
\value{
A flattened QTable, where a higher dimensional array representation is flattened
        to a 2d (matrix) representation (or 3d with the table contains multiple statistics
        and the statistics are not dropped to the first statistic). If no flattening is
        required then the original table is returned. If flattening is required and the table
        has spans, then the resulting matrix has those spans incorporated into the row and
        column names.
}
\description{
Flatten a QTable
}
