% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{Sum}
\alias{Sum}
\title{General-Purpose Element-wise Summation}
\usage{
Sum(
  ...,
  remove.missing = TRUE,
  remove.rows = c("NET", "SUM", "Total"),
  remove.columns = c("NET", "SUM", "Total"),
  subset = NULL,
  weights = NULL,
  warn = FALSE
)
}
\arguments{
\item{...}{Objects to be summed; e.g. Q Tables, arrays}

\item{remove.missing}{Should missing values be removed in calculations? Defaults to \code{TRUE},}

\item{remove.rows}{The categories of named vectors, or, rows of tables and similar
structures to be removed.}

\item{remove.columns}{The labels of categories to be removed from the calculation.
Where the table is two-dimensional,
differing from \code{na.rm} in \code{\link{sum}}.}

\item{warn}{Warnings are shown when non-obvious operations are performed (e.g.,
removing rows).}
}
\value{
The sum of the provided inputs
}
\description{
A replacement for \code{\link{sum}} supporting element-wise
summation of many objects in addition to vectors.
}
\examples{
Sum(1:3, 2:4)
}
